completed OrderService completed

+++ src/main/java/com/example/E_commerce/Service/UserService.java
@@ -1,7 +1,11 @@
 package com.example.E_commerce.Service;
 
+import com.example.E_commerce.Model.Order;
 import com.example.E_commerce.Model.User;
+import com.example.E_commerce.Repository.OrderRepository;
+import com.example.E_commerce.Repository.TransactionRepository;
 import com.example.E_commerce.Repository.UserRepository;
+import jakarta.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -13,6 +17,12 @@ public class UserService {
 
     @Autowired
     private UserRepository userRepository;
+    @Autowired
+    private OrderRepository orderRepository;
+    @Autowired
+    private TransactionRepository transactionRepository;
+    @Autowired
+    private OrderService orderService;
 
     public User addUser(User user){ //checked
         return userRepository.save(user);
@@ -21,12 +31,18 @@ public class UserService {
     public List<User> GetAll(){ //checked
         return userRepository.findAll();
     }
-
-    public String Delete(int id){ //checked
-        User user = new User();
-        String name = user.getName();
-        userRepository.deleteById(id);
-        return name+" Has been removed!!";
+    @Transactional
+    public String Delete(int id){//checked
+        Optional<User> user = userRepository.findById(id);
+        if (user.isPresent()){
+            User user1 = user.get();
+            String name = user1.getName();
+            List<Order> orders = orderService.GetOrderUser(id);
+            userRepository.deleteById(id);
+            return name+" Has been removed!!";
+        }else{
+            return "User with id "+id+" not found.";
+        }
     }
 
     public String Update(int id,String newName,String newUsername,String newPassword){ //checked
